admin update

+++ team_project/src/main/webapp/admin/userList.jsp
@@ -9,13 +9,42 @@
 <html>
 <head>
 <meta charset="UTF-8">
+<link rel="stylesheet" href="css/admain.css" type="text/css" >
 <title>Insert title here</title>
 </head>
 <body>
 <!-- 헤더들어가는 곳 -->
 <jsp:include page="../inc/ad_header.jsp" />
 <!-- 헤더들어가는 곳 -->
-
+		
+<!-- Backgrounds -->
+<div class="menu-item">
+	<div class="container">
+    	<div class="row">
+        	<div class="col-lg-2">
+				<div class="sidenav">		
+					<button class="dropdown-btn">Q&A <i class="fa fa-caret-down"></i>
+					</button>
+				<div class="dropdown-container">
+					<a href="#">Link 1</a> <a href="#">Link 2</a> <a href="#">Link 3</a> 
+				</div>
+					<button class="dropdown-btn">공지사항 <i class="fa fa-caret-down"></i>
+					</button>
+				<div class="dropdown-container">
+					<a href="#">Link 1</a> <a href="#">Link 2</a> <a href="#">Link 3</a>
+				</div> 
+					<a href="AdminAppointNow.ad">숙소예약현황</a> 
+					<a href="AdminUserList.ad">유저관리</a>
+					<button class="dropdown-btn">기타 <i class="fa fa-caret-down"></i>
+					</button>
+				<div class="dropdown-container">
+					<a href="#">Link 1</a> <a href="#">Link 2</a> <a href="#">Link 3</a>
+				</div>	
+				</div>
+			</div>
+			
+<div class="col-lg-10">
+	<div class="content-main">			
 <%
 // UserDAO dao=new UserDAO();
 //한 페이지레 보여줄 글개수 설정
@@ -65,7 +94,7 @@ for(int i=0;i<userList.size();i++){
        사장님
     <% } %>
 	</td>
-	<td><a href="AdminUserDelete.ad?id=<%=dto.getId()%>">회원삭제</a></td></tr>
+	<td><a href="AdminUserDelete.ad?id=<%=dto.getId()%>" onclick="return confirm('회원을 삭제하시겠습니까?')">회원삭제</a></td></tr>
 	<%
 }
 %>
@@ -110,5 +139,10 @@ if(endPage < pageCount){
 %>
 </div>
 </form>
+</div>
+</div>
+</div>
+</div>
+</div>
 </body>
 </html>
